introduced helepr_onnx.py and removed dupicate definitions
