+++ onnxruntime/test/python/helper.py
@@ -1,20 +1,4 @@
 import os
-
-
-def get_name(name):
-    if os.path.exists(name):
-        return name
-    rel = os.path.join("testdata", name)
-    if os.path.exists(rel):
-        return rel
-    this = os.path.dirname(__file__)
-    data = os.path.join(this, "..", "testdata")
-    res = os.path.join(data, name)
-    if os.path.exists(res):
-        return res
-    raise FileNotFoundError(f"Unable to find '{name}' or '{rel}' or '{res}'")
-import os
-import onnx
 import numpy as np
 import json
 import pandas as pd
@@ -31,18 +15,6 @@ def get_name(name):
     if os.path.exists(res):
         return res
     raise FileNotFoundError(f"Unable to find '{name}' or '{rel}' or '{res}'")
-
-def get_onnx_const(name, val, dtype=None):
-    if isinstance(val, np.ndarray):
-        dtype = onnx.mapping.NP_TYPE_TO_TENSOR_TYPE[val.dtype]
-        dims = val.shape
-    else:
-        if not dtype:
-            dtype = onnx.TensorProto.INT8 if isinstance(val, int) else onnx.TensorProto.FLOAT
-        dims = ()
-        val = [val]
-
-    return onnx.helper.make_tensor(name=name, data_type=dtype, dims=dims, vals=val)
 
 def generate_normal_inputs(shape, dtype, mu=0, sigma=32, a_min=-127, a_max=127):
     return np.clip(np.rint(np.random.normal(mu, sigma, shape)).astype(dtype), a_min, a_max)
